requirements.txt
torch==2.10.0
torchvision==0.25.0
torchinfo==1.8.0
tqdm==4.67.2
streamlit==1.54.0
scikit-learn==1.8.0
pycocotools==2.0.11
pillow==12.1.0
pandas==2.3.3
numpy==2.4.2
matplotlib==3.10.8
grad-cam==1.5.5